Refactor SongView and BaseView
